--- a/benchmark/benchmark_utils.jl
+++ b/benchmark/benchmark_utils.jl
@@ -1,3 +1,5 @@
+using Oceananigans: AbstractArchitecture
+
 arch_name(::CPU) = "CPU"
 arch_name(::GPU) = "GPU"