--- a/src/Oceananigans.jl
+++ b/src/Oceananigans.jl
@@ -247,6 +247,7 @@ include("buoyancy.jl")
 include("boundary_conditions.jl")
 include("halo_regions.jl")
 include("poisson_solvers.jl")
+include("model_initialization.jl")
 include("models.jl")
 include("time_steppers.jl")