--- a/src/Oceananigans.jl
+++ b/src/Oceananigans.jl
@@ -234,6 +234,7 @@ function buoyancy_perturbation end
 function buoyancy_frequency_squared end
 
 include("utils.jl")
+include("model_initialization_utils.jl")
 
 include("clock.jl")
 include("grids.jl")
@@ -247,7 +248,6 @@ include("buoyancy.jl")
 include("boundary_conditions.jl")
 include("halo_regions.jl")
 include("poisson_solvers.jl")
-include("model_initialization.jl")
 include("models.jl")
 include("time_steppers.jl")