--- a/src/Oceananigans.jl
+++ b/src/Oceananigans.jl
@@ -229,9 +229,10 @@ end
 architecture(::Array) = CPU()
 @hascuda architecture(::CuArray) = GPU()
 
-# Place-holder buoyancy functions for use in TurbulenceClosures module
+# Place-holder functions for use in TurbulenceClosures module
 function buoyancy_perturbation end
 function buoyancy_frequency_squared end
+function TracerFields end
 
 include("utils.jl")