--- a/src/TurbulenceClosures/TurbulenceClosures.jl
+++ b/src/TurbulenceClosures/TurbulenceClosures.jl
@@ -109,8 +109,8 @@ include("velocity_tracer_gradients.jl")
 include("constant_isotropic_diffusivity.jl")
 include("constant_anisotropic_diffusivity.jl")
 include("smagorinsky.jl")
-include("rozema_anisotropic_minimum_dissipation.jl")
 include("verstappen_anisotropic_minimum_dissipation.jl")
+include("rozema_anisotropic_minimum_dissipation.jl")
 
 include("turbulence_closure_diagnostics.jl")