--- a/src/TurbulenceClosures/constant_isotropic_diffusivity.jl
+++ b/src/TurbulenceClosures/constant_isotropic_diffusivity.jl
@@ -37,10 +37,10 @@ function with_tracers(tracers, closure::ConstantIsotropicDiffusivity{FT}) where 
     return ConstantIsotropicDiffusivity{FT}(closure.ν, κ)
 end
 
-calc_diffusivities!(diffusivities, grid, closure::ConstantIsotropicDiffusivity, args...) = nothing
+calc_diffusivities!(K, arch, grid, closure::ConstantIsotropicDiffusivity, args...) = nothing
 
-@inline function ∇_κ_∇c(i, j, k, grid, c, tracer, closure::ConstantIsotropicDiffusivity, args...)
-    κ = getproperty(closure.κ, tracer) 
+@inline function ∇_κ_∇c(i, j, k, grid, c, tracer_idx, closure::ConstantIsotropicDiffusivity, args...)
+    @inbounds κ = closure.κ[tracer_idx]
 
     return (  κ / grid.Δx^2 * δx²_c2f2c(grid, c, i, j, k)
             + κ / grid.Δy^2 * δy²_c2f2c(grid, c, i, j, k)