--- a/src/TurbulenceClosures/rozema_anisotropic_minimum_dissipation.jl
+++ b/src/TurbulenceClosures/rozema_anisotropic_minimum_dissipation.jl
@@ -63,7 +63,7 @@ end
     return max(zero(FT), νˢᵍˢ) + closure.ν
 end
 
-@inline function κ_ccc(i, j, k, grid::AbstractGrid{FT}, closure::RAMD, c, tracer_idx, buoyancy, U) where FT
+@inline function κ_ccc(i, j, k, grid::AbstractGrid{FT}, closure::RAMD, c, tracer_idx, U) where FT
     @inbounds κ = closure.κ[tracer_idx]
 
     σ = θᵢ²_ccc(i, j, k, grid, c)