This PR introduces 'arbitrary tracer' functionality to Oceananigans.

With this PR, the user is now able to specify which tracers they would like to include in their simulation via the `Model` constructor:

```julia
model = Model(..., tracers=(:b, :c1, :c2), ...)
```

## User-facing considerations

1. The specified tracers must be consistent with the specified buoyancy model. This consistency is ensured by `validate_buoyancy` in the `Model` constructor. For `BuoyancyTracer`, the user must specify a tracer named `:b`. For `SeawaterBuoyancy`, the user must specify tracers named `:T` and `:S`. (In the future, we could also allow the names of these buoyancy, temperature, and salinity tracers to be specified in the constructor for the respective buoyancy models, eg `SeawaterBuoyancy(temperature=:T, salinity=:S)`.)

2. The constructor for `ModelForcing` (which will have to be merged with #444 when it is approved) now allows the user to specify forcing for any of the tracer fields. If the user specifies a forcing for a tracer field that does not exist, an error is thrown. If a user does not specify a forcing for some tracer field, a default zero forcing is applied.

3. The same logic for `ModelForcing` applies to `BoundaryConditions` / `SolutionBoundaryConditions`. When boundary conditions for a tracer field are unspecified, a default tracer boundary condition is inferred from the x-velocity field: tracers either inherit `Periodic` boundary conditions, or are given a `Flux, Nothing` boundary condition for all other cases.

4. For closures, all tracer-diffusivity-related fields permit users to specify either:

* a constant (applied to all tracers), or 
* a `NamedTuple` with fields for *every* tracer. 

In the future, we could possibly implement some mixed behavior where the user specifies both a default *and* particular values for tracer diffusivity. This would be useful in the (possibly rare) use-case of a large number of tracers with the same diffusivity, but where one or two of them require a special, different diffusivity. I am not sure this API is necessary so I left it for future work.

Diffusivity-like fields include:

* `κ` (constant component of tracer diffusivity) for `ConstantIsotropicDiffusivity`, `AnisotropicMinimumDissipation`, and `ConstantSmagorinsky`
* `Pr` (turbulent Prandtl number) for `ConstantSmagorinsky`
* `κh` and `κv` for `ConstantAnisotropicDiffusivity`.

## Internal algorithmic considerations

This implementation includes a major refactor of the time-stepping algorithm. In particular, kernels are launched for *each* tracer for all operations that involve tracers. This differs from the previous pattern, in which a single kernel was called in some cases (for example, to update the velocity and tracer fields, or to store previous source terms). The reason for this change is because I ran into some issues (dynamic function invocations) using `ntuple` to unroll a loop over tracers inside the kernel. In addition, I think that with a large number of tracers the kernels may become too large and their performance could degrade (but I'm not sure).

This refactoring of the algorithm means we need to 

- [ ] benchmark the changes in this PR to see if there are any significant changes in model performance. 

@ali-ramadhan, can you help with this?

If there are changes in model performance, we can work on unrolling loops over the tracer fields *inside* our kernels. This is probably possible; it just requires some debugging. We would probably also want to make sure that this doesn't lead to poor performance for up to O(10) tracers.

If any of this PR is not satisfactory, I'm happy to work on it and iterate until the PR is in mergeable shape.

Resolves #25 
Resolves #430 
Partially addresses #448