@glwagner 2019-10-10
There is now a branch called `arbitrary-tracers-inner-loops` with an alternative implementation of arbitrary tracers against which benchmarks should be compared.