@ali-ramadhan 2019-10-12
Just playing around with benchmarks before I review this PR but I feel like this should error. I can probably fix when I review.

```julia
julia> model = Model(architecture=CPU(), grid=RegularCartesianGrid(N=(64, 64, 64), L=(1, 1, 1)), buoyancy=BuoyancyTracer);

julia> propertynames(model.tracers)
(:T, :S)
```