@glwagner 2019-10-12
> Just playing around with benchmarks before I review this PR but I feel like this should error.

Make sure you are using the right version of `Oceananigans`