@ali-ramadhan 2019-10-12
Pretty sure there's a performance regression on GPUs.

128^3 `Float64` GPU model slowed down by ~2.4x and 256^3 by ~2.9x while 64^3 didn't really slow down so maybe it's something that gets worse as the model size grows? Although 64^3 doesn't saturate the GPU.

I'll think about this as I review the PR...

---

### Master
 ──────────────────────────────────────────────────────────────────────────────────────
        Static ocean benchmarks                Time                   Allocations      
                                       ──────────────────────   ───────────────────────
           Tot / % measured:                 171s / 38.5%           14.9GiB / 0.44%    

 Section                       ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────────────
 256×256×256  (CPU, Float32)       10    29.2s  44.4%   2.92s    733KiB  1.06%  73.3KiB
 256×256×256  (CPU, Float64)       10    28.8s  43.9%   2.88s    733KiB  1.06%  73.3KiB
 128×128×128  (CPU, Float32)       10    3.33s  5.07%   333ms    733KiB  1.06%  73.3KiB
 128×128×128  (CPU, Float64)       10    2.87s  4.36%   287ms    733KiB  1.06%  73.3KiB
  64× 64× 64  (CPU, Float32)       10    378ms  0.58%  37.8ms    733KiB  1.06%  73.3KiB
  64× 64× 64  (CPU, Float64)       10    314ms  0.48%  31.4ms    733KiB  1.06%  73.3KiB
 256×256×256  (GPU, Float32)       10    278ms  0.42%  27.8ms   7.71MiB  11.4%   789KiB
 256×256×256  (GPU, Float64)       10    270ms  0.41%  27.0ms   7.73MiB  11.4%   791KiB
 128×128×128  (GPU, Float64)       10   45.4ms  0.07%  4.54ms   7.73MiB  11.4%   791KiB
 128×128×128  (GPU, Float32)       10   43.1ms  0.07%  4.31ms   7.71MiB  11.4%   789KiB
  32× 32× 32  (CPU, Float32)       10   41.0ms  0.06%  4.10ms    733KiB  1.06%  73.3KiB
  32× 32× 32  (CPU, Float64)       10   41.0ms  0.06%  4.10ms    733KiB  1.06%  73.3KiB
  64× 64× 64  (GPU, Float64)       10   36.2ms  0.06%  3.62ms   7.84MiB  11.6%   803KiB
  64× 64× 64  (GPU, Float32)       10   36.2ms  0.06%  3.62ms   7.82MiB  11.6%   801KiB
  32× 32× 32  (GPU, Float32)       10   33.2ms  0.05%  3.32ms   7.70MiB  11.4%   789KiB
  32× 32× 32  (GPU, Float64)       10   32.2ms  0.05%  3.22ms   7.72MiB  11.4%   791KiB
 ──────────────────────────────────────────────────────────────────────────────────────


### This PR
 ──────────────────────────────────────────────────────────────────────────────────────
        Static ocean benchmarks                Time                   Allocations      
                                       ──────────────────────   ───────────────────────
           Tot / % measured:                 176s / 38.7%           15.2GiB / 0.51%    

 Section                       ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────────────
 256×256×256  (CPU, Float64)       10    30.0s  44.1%   3.00s    725KiB  0.90%  72.5KiB
 256×256×256  (CPU, Float32)       10    29.3s  42.9%   2.93s    725KiB  0.90%  72.5KiB
 128×128×128  (CPU, Float32)       10    3.26s  4.78%   326ms    725KiB  0.90%  72.5KiB
 128×128×128  (CPU, Float64)       10    2.93s  4.30%   293ms    725KiB  0.90%  72.5KiB
 256×256×256  (GPU, Float64)       10    780ms  1.14%  78.0ms   9.16MiB  11.6%   938KiB
 256×256×256  (GPU, Float32)       10    779ms  1.14%  77.9ms   9.16MiB  11.6%   938KiB
  64× 64× 64  (CPU, Float32)       10    351ms  0.51%  35.1ms    725KiB  0.90%  72.5KiB
  64× 64× 64  (CPU, Float64)       10    325ms  0.48%  32.5ms    725KiB  0.90%  72.5KiB
 128×128×128  (GPU, Float64)       10    107ms  0.16%  10.7ms   9.16MiB  11.6%   938KiB
 128×128×128  (GPU, Float32)       10    106ms  0.16%  10.6ms   9.16MiB  11.6%   938KiB
  32× 32× 32  (CPU, Float64)       10   43.4ms  0.06%  4.34ms    725KiB  0.90%  72.5KiB
  32× 32× 32  (CPU, Float32)       10   42.2ms  0.06%  4.22ms    725KiB  0.90%  72.5KiB
  32× 32× 32  (GPU, Float32)       10   34.4ms  0.05%  3.44ms   9.16MiB  11.6%   938KiB
  64× 64× 64  (GPU, Float64)       10   34.2ms  0.05%  3.42ms   9.16MiB  11.6%   938KiB
  64× 64× 64  (GPU, Float32)       10   33.8ms  0.05%  3.38ms   9.16MiB  11.6%   938KiB
  32× 32× 32  (GPU, Float64)       10   33.1ms  0.05%  3.31ms   9.16MiB  11.6%   938KiB
 ──────────────────────────────────────────────────────────────────────────────────────