@glwagner 2019-10-12
Unrolling inside loops may be the way to go then.