@glwagner 2019-10-15
> Out of curiousity, would a user ever want to run some tracers with LES and other tracers with constant diffusivity? Seems like it'll be a weird/inconsistent model.

That's not out of the question. We could make the `closure` itself a tuple for different tracers. This functionality could also be hacked together for smagorinsky if `Pr = Inf` for one or more tracers. However, we don't flexibly permit the mixing and matching of different closures in general. Something to ponder for the future, in the general category of how we allow users to more flexibly change the equation they are solving.